feat: set login logic when first come in(have token) - split rest auth <-> login auth
TODO
1. have to union it
