+++ srcs/auth/index.tsx
@@ -1,0 +1,3 @@
+export { fail, login, granted } from './controller.login';
+export { default as signup } from './controller.signup';
+export { default as isAuth } from './auth';
